auth

+++ backend/app/Http/Controllers/PostController.php
@@ -4,11 +4,21 @@ namespace App\Http\Controllers;
 
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 
 use function Laravel\Prompts\search;
 
-class PostController extends Controller
+class PostController extends Controller implements HasMiddleware
 {
+    public static function middleware()
+    {
+        return [
+            new Middleware('auth:sanctum', except: ['index', 'show'])
+        ];
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -27,13 +37,10 @@ class PostController extends Controller
             'body' => 'required'
         ]);
 
-        $post = Post::create($fields);
+        $post = $request->user()->posts()->create($fields);
 
-        // return ['post' => $post];
         return response()->json([
-            'status' => 200,
-            'message' => 'Post created',
-            'post' => $post
+            'message' => 'Post created'
         ], 200);
     }
 
@@ -60,6 +67,8 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        Gate::authorize('modify', $post);
+
         $fields = $request->validate([
             'title' => 'required|max:255',
             'body' => 'required'
@@ -78,7 +87,9 @@ class PostController extends Controller
      * Remove the specified resource from storage.
      */
     public function destroy(Post $post)
-    {
+    {   
+        Gate::authorize('modify', $post);
+
         $post->delete();
 
         return response()->json([
